+++ api/src/routes/index.mjs
@@ -1,9 +1,10 @@
-import { routes as anonymous } from '#routes/anonymous'
-import { routes as core } from '#routes/core'
-import { routes as auth } from '#routes/auth'
 import { routes as accounts } from '#routes/accounts'
+import { routes as anonymous } from '#routes/anonymous'
 import { routes as apikeys } from '#routes/apikeys'
+import { routes as auth } from '#routes/auth'
+import { routes as cache } from '#routes/cache'
 import { routes as crypto } from '#routes/crypto'
+import { routes as core } from '#routes/core'
 import { routes as docs } from '#routes/docs'
 import { routes as kv } from '#routes/kv'
 
@@ -12,6 +13,7 @@ export const routes = {
   auth,
   accounts,
   apikeys,
+  cache,
   core,
   crypto,
   docs,